src/requirements.txt
asn1crypto==0.22.0
Babel==2.3.4
cffi==1.9.1
coreapi==2.2.4
coreschema==0.0.4
cryptography==1.8.1
dj-database-url==0.4.2
Django==1.10.5
django-annoying==0.10.3
django-appconf==1.0.2
django-braces==1.11.0
django-colorfield==0.1.12
django-coreapi==0.5.0
django-countries==4.1
django-crispy-forms==1.6.1
django-extensions==1.7.7
django-formtools==2.0
django-guardian==1.4.6
django-markdown-app==0.9.0
django-model-utils==2.6.1
django-phonenumber-field==1.1.0
django-timezone-field==2.0
django-tinymce==2.6.0
djangorestframework==3.5.4
graphviz==0.6
gunicorn==19.7.0
icalendar==3.11.3
idna==2.5
itypes==1.1.0
Jinja2==2.9.5
jsonfield==2.0.1
Markdown==2.6.8
MarkupSafe==0.23
ndg-httpsclient==0.4.2
packaging==16.8
phonenumberslite==8.3.2
psycopg2==2.7
pyasn1==0.2.3
pycparser==2.17
pyOpenSSL==16.2.0
pyparsing==2.1.10
python-dateutil==2.6.0
pytz==2016.10
requests==2.13.0
six==1.10.0
stripe==1.50.0
uritemplate==3.0.0
urllib3==1.20
whitenoise==3.3.0
